reactor: Hashtag Select ViewModel 분리(회원가입에서 사용되는 ViewModel)
