+++ Mogakco/Sources/Presentation/Common/Coordinator/AppCoordinator.swift
@@ -21,13 +21,7 @@ final class AppCoordinator: Coordinator, AppCoordinatorProtocol {
     }
 
     func start() {
-        let hashtagDataSource = HashtagDataSource()
-        let hashtagRepository = HashtagRepository(localHashtagDataSource: hashtagDataSource)
-        let hashtagUsecase = HashtagUsecase(hashtagRepository: hashtagRepository)
-        let viewModel = HashtagSelectViewModel(coordinator: AdditionalSignupCoordinator(navigationController), hashTagUsecase: hashtagUsecase)
-        let hashtagSelectViewController = HashtagSelectViewController(kind: .career, viewModel: viewModel)
-        navigationController.pushViewController(hashtagSelectViewController, animated: true)
-//        showAuthFlow()
+        showAuthFlow()
     }
     
     func showAuthFlow() {
