+++ Mogakco/Sources/Presentation/Hashtag/ViewModel/HashtagSelectViewModel.swift
@@ -16,95 +16,49 @@ protocol Hashtag {
     func hashtagTitle() -> String
 }
 
-class HashtagSelectViewModel: ViewModel {
+class HashtagSelectedViewModel: HashtagViewModel {
     
-    struct Input {
-        let kindHashtag: Observable<KindHashtag>
-        let cellSelected: Observable<Int>
-        let nextButtonTapped: Observable<Void>
-    }
-    struct Output {
-        let hashtagReload: Observable<Void>
-    }
-    
-    weak var coordinator: AdditionalSignupCoordinatorProtocol?
-    let hashTagUsecase: HashtagUseCaseProtocol
-    var disposeBag = DisposeBag()
-    var selectedHashtag: [Hashtag] = []
-    let badgeList = BehaviorSubject<[Hashtag]>(value: [])
-    private var kind: KindHashtag = .language
-    
-    var collectionViewCount: Int {
-        guard let count = try? badgeList.value().count else { return 0 }
-        return count
-    }
+    let signUseCase: SignupUseCaseProtocol?
     
     init(
         coordinator: AdditionalSignupCoordinatorProtocol,
-        hashTagUsecase: HashtagUseCaseProtocol
+        hashTagUsecase: HashtagUseCaseProtocol,
+        signUpUseCase: SignupUseCaseProtocol? = nil
     ) {
-        self.coordinator = coordinator
-        self.hashTagUsecase = hashTagUsecase
+        self.signUseCase = signUpUseCase
+        super.init(coordinator: coordinator, hashTagUsecase: hashTagUsecase)
     }
     
-    func transform(input: Input) -> Output {
-        let collectionReloadObservable = PublishSubject<Void>()
-        
-        input.kindHashtag
-            .subscribe { [weak self] in
-                self?.loadTagList(kind: $0)
-                self?.kind = $0
-            }
-            .disposed(by: disposeBag)
-        
-        input.cellSelected
-            .withUnretained(self)
-            .subscribe { owner, index in
-                owner.selectHashtag(index: index)
-            }
-            .disposed(by: disposeBag)
+    override func transform(input: Input) -> Output {
         
         input.nextButtonTapped
-            .subscribe(onNext: {
-                // TODO: 분기처리 필요
-            })
-            .disposed(by: disposeBag)
-        
-        badgeList
-            .subscribe { _ in
-                collectionReloadObservable.onNext(())
-            }
-            .disposed(by: disposeBag)
-        
-        return Output(hashtagReload: collectionReloadObservable.asObservable())
-    }
-    
-    func cellInfo(index: Int) -> Hashtag? {
-        return try? badgeList.value()[index]
-    }
-    
-    func isSelected(index: Int) -> Bool {
-        guard let hashtag = cellInfo(index: index) else { return false }
-        if selectedHashtag.contains(where: { $0.title == hashtag.title }) { return true }
-        
-        return false
-    }
-    
-    private func loadTagList(kind: KindHashtag) {
-        hashTagUsecase.loadTagList(kind: kind)
             .subscribe { [weak self] in
-                self?.badgeList.onNext($0)
+                self?.moveToNext()
             }
             .disposed(by: disposeBag)
+        
+        return super.transform(input: input)
     }
     
-    private func selectHashtag(index: Int) {
-        guard let hashTag = cellInfo(index: index) else { return }
-        
-        if let removeIndex = selectedHashtag.firstIndex(where: { $0.title == hashTag.title }) {
-            selectedHashtag.remove(at: removeIndex)
-        } else {
-            selectedHashtag.append(hashTag)
+    private func moveToNext() {
+        switch kind {
+        case .language: coordinator?.showCareer() // TODO: 정보 전달 로직 필요
+        case .career: signUseCase?.signup(
+            user: User(
+                id: "",
+                email: "",
+                introduce: "",
+                password: "",
+                name: "",
+                languages: [],
+                careers: [],
+                categorys: []
+            )
+        ).subscribe { _ in
+            // 회원가입 후 이동
+        }
+        .disposed(by: disposeBag)
+        case .category: break
         }
     }
 }
